docs: Reorganise stories internally to separate NRTs from core doc
