+++ src/examples/Badge.stories.ts
@@ -1,4 +1,3 @@
-import React from 'react'
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
 import { Badge } from '../components/Badge'
@@ -114,32 +113,4 @@ export const SidebarBadge: Story = {
     text: 'Text',
     tooltip: 'Tooltips are disabled in the sidebar',
   },
-}
-
-export const TightSpace: Story = {
-  args: {
-    context: 'sidebar',
-    text: 'Multi-word Badge',
-  },
-  tags: ['very-tight-space'],
-  decorators: [
-    (story) => (
-      <div style={{ width: '60px', background: 'orange', overflow: 'hidden' }}>
-        {story()}
-      </div>
-    ),
-  ],
-}
-
-export const VeryTightSpaceWithNoSpaceForABadge: Story = {
-  args: {
-    context: 'sidebar',
-    text: 'Multi-word Badge',
-  },
-  tags: ['very-tight-space'],
-  decorators: [
-    (story) => (
-      <div style={{ width: '20px', background: 'red' }}>{story()}</div>
-    ),
-  ],
 }
