+++ src/stories/Issue86.stories.ts
@@ -1,15 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
-import { fn } from 'storybook/test'
 
 import { Button } from './Button'
 
 const meta: Meta<typeof Button> = {
   title: 'NRT/issue-86/Button',
   component: Button,
-  argTypes: {
-    backgroundColor: { control: 'color' },
-    onClick: fn(),
-  },
   tags: ['dd-privacy:allow'],
 }
